Refactor amount and fee rate types from ulong to long across multiple models and view models

+++ src/Angor/Avalonia/Angor.Model.Implementation/Projects/Stage.cs
@@ -5,7 +5,7 @@ namespace Angor.Model.Implementation.Projects;
 public class Stage : IStage
 {
     public DateOnly ReleaseDate { get; set; }
-    public uint Amount { get; set; }
+    public long Amount { get; set; }
     public int Index { get; set; }
     public double Weight { get; set; }
 }
