+++ src/Angor/Avalonia/Angor.Model/IStage.cs
@@ -3,7 +3,7 @@ namespace AngorApp.Model;
 public interface IStage
 {
     DateOnly ReleaseDate { get; }
-    uint Amount { get; }
+    long Amount { get; }
     int Index { get; }
     double Weight { get; }
 }
