+++ src/Angor/Avalonia/AngorApp/Sections/Browse/StageDesign.cs
@@ -5,7 +5,7 @@ namespace AngorApp.Sections.Browse;
 public class StageDesign : IStage
 {
     public DateOnly ReleaseDate { get; set; }
-    public uint Amount { get; set; }
+    public long Amount { get; set; }
     public int Index { get; set; }
     public double Weight { get; set; }
 }
